Standardize the SCOs height
A method with 30 NOS had an other building height than the building of a
report in the old procession. The reason was the old calculation with
building parts, but this is in the SimpleBlocks mode not necessary. Now
the buildings height of a SCO is independent of the type. Further there
is a scale factor so that the buildings with height = 1 are not too
inconspicuous.